+++ ios/BLEButtonManager.m
@@ -9,6 +9,10 @@ RCT_EXTERN_METHOD(connectToDevice:(NSString *)uuid)
 RCT_EXTERN_METHOD(disconnectDevice)
 RCT_EXTERN_METHOD(getSavedDeviceUUID:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(getSavedDevices:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(forgetDevice:(NSString *)uuid)
+RCT_EXTERN_METHOD(connectToAnySaved)
 RCT_EXTERN_METHOD(getState:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 RCT_EXTERN_METHOD(probeDevice)